+++ beaker/services/organization.py
@@ -10,11 +10,11 @@ class OrganizationClient(ServiceClient):
     Accessed via :data:`Beaker.organization <beaker.Beaker.organization>`.
     """
 
-    def get(self, name: Optional[str] = None) -> Organization:
+    def get(self, org: Optional[str] = None) -> Organization:
         """
         Get information about an organization.
 
-        :param name: The organization name. If not specified,
+        :param org: The organization name or ID. If not specified,
             :data:`Beaker.config.default_org <beaker.Config.default_org>` is used.
 
         :raises OrganizationNotFound: If the organization doesn't exist.
@@ -22,12 +22,15 @@ class OrganizationClient(ServiceClient):
             :data:`Beaker.config.default_org <beaker.Config.default_org>` are set.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        name = self._resolve_org_name(name)
+        org = org or self.config.default_org
+        if org is None:
+            raise OrganizationNotSet("'org' argument required since default org not set")
+
         return Organization.from_json(
             self.request(
-                f"orgs/{self._url_quote(name)}",
+                f"orgs/{self.url_quote(org)}",
                 method="GET",
-                exceptions_for_status={404: OrganizationNotFound(name)},
+                exceptions_for_status={404: OrganizationNotFound(org)},
             ).json()
         )
 
@@ -47,10 +50,10 @@ class OrganizationClient(ServiceClient):
         :raises AccountNotFound: If the account doesn't exist.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        org: Organization = self._resolve_org(org)
+        org: Organization = self.resolve_org(org)
         account_name = account if isinstance(account, str) else account.name
         self.request(
-            f"orgs/{self._url_quote(org.name)}/members/{account_name}",
+            f"orgs/{self.url_quote(org.name)}/members/{account_name}",
             method="PUT",
             exceptions_for_status={404: AccountNotFound(account_name)},
         )
@@ -72,11 +75,11 @@ class OrganizationClient(ServiceClient):
         :raises AccountNotFound: If the account doesn't exist or isn't a member of the org.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        org: Organization = self._resolve_org(org)
+        org: Organization = self.resolve_org(org)
         account_name = account if isinstance(account, str) else account.name
         return OrganizationMember.from_json(
             self.request(
-                f"orgs/{self._url_quote(org.name)}/members/{account_name}",
+                f"orgs/{self.url_quote(org.name)}/members/{account_name}",
                 method="GET",
                 exceptions_for_status={404: AccountNotFound(account_name)},
             ).json()
@@ -94,11 +97,11 @@ class OrganizationClient(ServiceClient):
             :data:`Beaker.config.default_org <beaker.Config.default_org>` are set.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        org: Organization = self._resolve_org(org)
+        org: Organization = self.resolve_org(org)
         return [
             Account.from_json(d)
             for d in self.request(
-                f"orgs/{self._url_quote(org.name)}/members",
+                f"orgs/{self.url_quote(org.name)}/members",
                 method="GET",
                 exceptions_for_status={404: OrganizationNotFound(org.name)},
             ).json()["data"]
@@ -119,10 +122,10 @@ class OrganizationClient(ServiceClient):
             :data:`Beaker.config.default_org <beaker.Config.default_org>` are set.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        org: Organization = self._resolve_org(org)
+        org: Organization = self.resolve_org(org)
         account_name = account if isinstance(account, str) else account.name
         self.request(
-            f"orgs/{self._url_quote(org.name)}/members/{account_name}",
+            f"orgs/{self.url_quote(org.name)}/members/{account_name}",
             method="DELETE",
             exceptions_for_status={404: AccountNotFound(account_name)},
         )